Split minimap related code to source since it's getting bigger
